pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crash-reporter-backend"
version = "0.1.0"
description = "Crash Reporter backend API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "beanie>=1.26.0",
  "motor>=3.4.0",
  "pydantic>=2.8.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib>=1.7.4",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "mongomock-motor>=0.0.35",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
